Conflicts:
	Erebus/Scripts/level.lua
	Erebus/Scripts/player.lua

+++ Erebus/LuaInputs.cpp
@@ -66,7 +66,7 @@ namespace LuaInputs
 			result = str[0];
 		}
 		else
-			result = lua_tointeger( lua, 1 );
+			result = (int)lua_tointeger( lua, 1 );
 
 		return result;
 	}
@@ -111,7 +111,7 @@ namespace LuaInputs
 	{
 		assert( lua_gettop( lua ) >= 1 );
 
-		unsigned int button = lua_tointeger( lua, 1 );
+		unsigned int button = (unsigned int)lua_tointeger( lua, 1 );
 		lua_pushboolean( lua, g_inputs->buttonPressed( button ) );
 		return 1;
 	}
@@ -120,7 +120,7 @@ namespace LuaInputs
 	{
 		assert( lua_gettop( lua ) >= 1 );
 
-		unsigned int button = lua_tointeger( lua, 1 );
+		unsigned int button = (unsigned int)lua_tointeger( lua, 1 );
 		lua_pushboolean( lua, g_inputs->buttonPressedThisFrame( button ) );
 		return 1;
 	}
@@ -129,7 +129,7 @@ namespace LuaInputs
 	{
 		assert( lua_gettop( lua ) >= 1 );
 
-		unsigned int button = lua_tointeger( lua, 1 );
+		unsigned int button = (unsigned int)lua_tointeger( lua, 1 );
 		lua_pushboolean( lua, g_inputs->buttonReleasedThisFrame( button ) );
 		return 1;
 	}
